\name{BIC_cdfpen}
\alias{BIC_cdfpen}
\title{BIC computation from a "cdfpen" object
}
\description{
Calculates the BIC for all estimated models in a "cdfpen" object}

\usage{
BIC_cdfpen(object)
}
\arguments{
  \item{object}{Object containing the results.}
}

\value{Returns a vector containing the BIC values calculated over the entire estimated path}
