\name{plot_path}
\alias{plot_path}
\title{Plotter function for cdfpen class
}
\description{
Function that takes user requests as input, to show the requested graph
}
\usage{
plot_path(obj,
          lmb,
          coeff,
          type = c("path", "l1", "BIC"),
          ...)
}
\arguments{
  \item{obj}{Object to be plotted
}
  \item{lmb}{lambda values used in the model
}
  \item{coeff}{the coefficients' matrix
}
  \item{type}{type of graph to be ploted
}
  \item{\dots}{Other characteristics to be added
}
}

\value{No return value}

