\name{S}
\alias{S}

\title{Threshold function for CDF penalty

}
\description{
Applies the threshold rule to obtain the vector of sparse estimates}
\usage{
S(bm_gm,
  db,
  w)
}

\arguments{
  \item{bm_gm}{Vector of pseudo-solution.}
  \item{db}{Lambda-rho ratio.}
  \item{w}{Weights obtained from the penalty function.}
}


\value{
The estimated coefficient
}
